+++ backend/internal/models/file_tree.go
@@ -13,6 +13,6 @@ type FileStatus struct {
 }
 
 type LineRange struct {
-	Start int `json:"start"`
-	End   int `json:"end"`
+	Start int64 `json:"start"`
+	End   int64 `json:"end"`
 }
