+++ backend/internal/models/regrade.go
@@ -3,5 +3,5 @@ package models
 type Regrade struct {
 	StudentGHUsername string `json:"student_gh_username" db:"student_gh_username"`
 	TAGHUsername      string `json:"ta_gh_username" db:"ta_gh_username"`
-	DueDateID         int32  `json:"due_date_id"`
+	DueDateID         int64  `json:"due_date_id"`
 }
